Create Closest Neighbour in BST-- march 31
Given the root of a binary search tree and a number n, find the greatest number in the binary search tree that is less than or equal to n. 

Your task :

You don't need to read input or print anything. Your task is to complete the function findMaxForN() which takes the root of the BST, and the integer n as input parameters and returns the greatest element less than or equal to n in the given BST, Return -1 if no such element exists.