Update 13-04-2024
Reverse Bits

Given a number x, reverse its binary form and return the answer in decimal.
Your Task:
You don't need to read input or print anything. Your task is to complete the function reversedBits() which takes an Integer x as input and returns the reverse binary form of x in decimal form.